split top-level functions defined in ExhaustiveDeps into a `utils` module
