+++ packages/eslint-plugin-react-hooks/src/ExhaustiveDeps.js
@@ -9,6 +9,17 @@
 
 'use strict';
 
+import {
+  collectRecommendations,
+  fastFindReferenceWithParent,
+  getDependency,
+  getReactiveHookCallbackIndex,
+  getNodeWithoutReactNamespace,
+  joinEnglish,
+  scanForDeclaredBareFunctions,
+  toPropertyAccessString,
+} from './utils';
+
 export default {
   meta: {
     type: 'suggestion',
@@ -1184,425 +1195,3 @@ export default {
     }
   },
 };
-
-// The meat of the logic.
-function collectRecommendations({
-  dependencies,
-  declaredDependencies,
-  optionalDependencies,
-  externalDependencies,
-  isEffect,
-}) {
-  // Our primary data structure.
-  // It is a logical representation of property chains:
-  // `props` -> `props.foo` -> `props.foo.bar` -> `props.foo.bar.baz`
-  //         -> `props.lol`
-  //         -> `props.huh` -> `props.huh.okay`
-  //         -> `props.wow`
-  // We'll use it to mark nodes that are *used* by the programmer,
-  // and the nodes that were *declared* as deps. Then we will
-  // traverse it to learn which deps are missing or unnecessary.
-  const depTree = createDepTree();
-  function createDepTree() {
-    return {
-      isRequired: false, // True if used in code
-      isSatisfiedRecursively: false, // True if specified in deps
-      hasRequiredNodesBelow: false, // True if something deeper is used by code
-      children: new Map(), // Nodes for properties
-    };
-  }
-
-  // Mark all required nodes first.
-  // Imagine exclamation marks next to each used deep property.
-  dependencies.forEach((_, key) => {
-    const node = getOrCreateNodeByPath(depTree, key);
-    node.isRequired = true;
-    markAllParentsByPath(depTree, key, parent => {
-      parent.hasRequiredNodesBelow = true;
-    });
-  });
-
-  // Mark all satisfied nodes.
-  // Imagine checkmarks next to each declared dependency.
-  declaredDependencies.forEach(({key}) => {
-    const node = getOrCreateNodeByPath(depTree, key);
-    node.isSatisfiedRecursively = true;
-  });
-  optionalDependencies.forEach(key => {
-    const node = getOrCreateNodeByPath(depTree, key);
-    node.isSatisfiedRecursively = true;
-  });
-
-  // Tree manipulation helpers.
-  function getOrCreateNodeByPath(rootNode, path) {
-    const keys = path.split('.');
-    let node = rootNode;
-    for (const key of keys) {
-      let child = node.children.get(key);
-      if (!child) {
-        child = createDepTree();
-        node.children.set(key, child);
-      }
-      node = child;
-    }
-    return node;
-  }
-  function markAllParentsByPath(rootNode, path, fn) {
-    const keys = path.split('.');
-    let node = rootNode;
-    for (const key of keys) {
-      const child = node.children.get(key);
-      if (!child) {
-        return;
-      }
-      fn(child);
-      node = child;
-    }
-  }
-
-  // Now we can learn which dependencies are missing or necessary.
-  const missingDependencies = new Set();
-  const satisfyingDependencies = new Set();
-  scanTreeRecursively(
-    depTree,
-    missingDependencies,
-    satisfyingDependencies,
-    key => key,
-  );
-  function scanTreeRecursively(node, missingPaths, satisfyingPaths, keyToPath) {
-    node.children.forEach((child, key) => {
-      const path = keyToPath(key);
-      if (child.isSatisfiedRecursively) {
-        if (child.hasRequiredNodesBelow) {
-          // Remember this dep actually satisfied something.
-          satisfyingPaths.add(path);
-        }
-        // It doesn't matter if there's something deeper.
-        // It would be transitively satisfied since we assume immutability.
-        // `props.foo` is enough if you read `props.foo.id`.
-        return;
-      }
-      if (child.isRequired) {
-        // Remember that no declared deps satisfied this node.
-        missingPaths.add(path);
-        // If we got here, nothing in its subtree was satisfied.
-        // No need to search further.
-        return;
-      }
-      scanTreeRecursively(
-        child,
-        missingPaths,
-        satisfyingPaths,
-        childKey => path + '.' + childKey,
-      );
-    });
-  }
-
-  // Collect suggestions in the order they were originally specified.
-  const suggestedDependencies = [];
-  const unnecessaryDependencies = new Set();
-  const duplicateDependencies = new Set();
-  declaredDependencies.forEach(({key}) => {
-    // Does this declared dep satisfy a real need?
-    if (satisfyingDependencies.has(key)) {
-      if (suggestedDependencies.indexOf(key) === -1) {
-        // Good one.
-        suggestedDependencies.push(key);
-      } else {
-        // Duplicate.
-        duplicateDependencies.add(key);
-      }
-    } else {
-      if (
-        isEffect &&
-        !key.endsWith('.current') &&
-        !externalDependencies.has(key)
-      ) {
-        // Effects are allowed extra "unnecessary" deps.
-        // Such as resetting scroll when ID changes.
-        // Consider them legit.
-        // The exception is ref.current which is always wrong.
-        if (suggestedDependencies.indexOf(key) === -1) {
-          suggestedDependencies.push(key);
-        }
-      } else {
-        // It's definitely not needed.
-        unnecessaryDependencies.add(key);
-      }
-    }
-  });
-
-  // Then add the missing ones at the end.
-  missingDependencies.forEach(key => {
-    suggestedDependencies.push(key);
-  });
-
-  return {
-    suggestedDependencies,
-    unnecessaryDependencies,
-    duplicateDependencies,
-    missingDependencies,
-  };
-}
-
-// Finds functions declared as dependencies
-// that would invalidate on every render.
-function scanForDeclaredBareFunctions({
-  declaredDependencies,
-  declaredDependenciesNode,
-  componentScope,
-  scope,
-}) {
-  const bareFunctions = declaredDependencies
-    .map(({key}) => {
-      const fnRef = componentScope.set.get(key);
-      if (fnRef == null) {
-        return null;
-      }
-      const fnNode = fnRef.defs[0];
-      if (fnNode == null) {
-        return null;
-      }
-      // const handleChange = function () {}
-      // const handleChange = () => {}
-      if (
-        fnNode.type === 'Variable' &&
-        fnNode.node.type === 'VariableDeclarator' &&
-        fnNode.node.init != null &&
-        (fnNode.node.init.type === 'ArrowFunctionExpression' ||
-          fnNode.node.init.type === 'FunctionExpression')
-      ) {
-        return fnRef;
-      }
-      // function handleChange() {}
-      if (
-        fnNode.type === 'FunctionName' &&
-        fnNode.node.type === 'FunctionDeclaration'
-      ) {
-        return fnRef;
-      }
-      return null;
-    })
-    .filter(Boolean);
-
-  function isUsedOutsideOfHook(fnRef) {
-    let foundWriteExpr = false;
-    for (let i = 0; i < fnRef.references.length; i++) {
-      const reference = fnRef.references[i];
-      if (reference.writeExpr) {
-        if (foundWriteExpr) {
-          // Two writes to the same function.
-          return true;
-        } else {
-          // Ignore first write as it's not usage.
-          foundWriteExpr = true;
-          continue;
-        }
-      }
-      let currentScope = reference.from;
-      while (currentScope !== scope && currentScope != null) {
-        currentScope = currentScope.upper;
-      }
-      if (currentScope !== scope) {
-        // This reference is outside the Hook callback.
-        // It can only be legit if it's the deps array.
-        if (!isAncestorNodeOf(declaredDependenciesNode, reference.identifier)) {
-          return true;
-        }
-      }
-    }
-    return false;
-  }
-
-  return bareFunctions.map(fnRef => ({
-    fn: fnRef.defs[0],
-    suggestUseCallback: isUsedOutsideOfHook(fnRef),
-  }));
-}
-
-/**
- * Assuming () means the passed/returned node:
- * (props) => (props)
- * props.(foo) => (props.foo)
- * props.foo.(bar) => (props).foo.bar
- * props.foo.bar.(baz) => (props).foo.bar.baz
- */
-function getDependency(node) {
-  if (
-    node.parent.type === 'MemberExpression' &&
-    node.parent.object === node &&
-    node.parent.property.name !== 'current' &&
-    !node.parent.computed &&
-    !(
-      node.parent.parent != null &&
-      node.parent.parent.type === 'CallExpression' &&
-      node.parent.parent.callee === node.parent
-    )
-  ) {
-    return getDependency(node.parent);
-  } else if (
-    node.type === 'MemberExpression' &&
-    node.parent &&
-    node.parent.type === 'AssignmentExpression'
-  ) {
-    return node.object;
-  } else {
-    return node;
-  }
-}
-
-/**
- * Assuming () means the passed node.
- * (foo) -> 'foo'
- * foo.(bar) -> 'foo.bar'
- * foo.bar.(baz) -> 'foo.bar.baz'
- * Otherwise throw.
- */
-function toPropertyAccessString(node) {
-  if (node.type === 'Identifier') {
-    return node.name;
-  } else if (node.type === 'MemberExpression' && !node.computed) {
-    const object = toPropertyAccessString(node.object);
-    const property = toPropertyAccessString(node.property);
-    return `${object}.${property}`;
-  } else {
-    throw new Error(`Unsupported node type: ${node.type}`);
-  }
-}
-
-function getNodeWithoutReactNamespace(node, options) {
-  if (
-    node.type === 'MemberExpression' &&
-    node.object.type === 'Identifier' &&
-    node.object.name === 'React' &&
-    node.property.type === 'Identifier' &&
-    !node.computed
-  ) {
-    return node.property;
-  }
-  return node;
-}
-
-// What's the index of callback that needs to be analyzed for a given Hook?
-// -1 if it's not a Hook we care about (e.g. useState).
-// 0 for useEffect/useMemo/useCallback(fn).
-// 1 for useImperativeHandle(ref, fn).
-// For additionally configured Hooks, assume that they're like useEffect (0).
-function getReactiveHookCallbackIndex(calleeNode, options) {
-  const node = getNodeWithoutReactNamespace(calleeNode);
-  if (node.type !== 'Identifier') {
-    return -1;
-  }
-  switch (node.name) {
-    case 'useEffect':
-    case 'useLayoutEffect':
-    case 'useCallback':
-    case 'useMemo':
-      // useEffect(fn)
-      return 0;
-    case 'useImperativeHandle':
-      // useImperativeHandle(ref, fn)
-      return 1;
-    default:
-      if (node === calleeNode && options && options.additionalHooks) {
-        // Allow the user to provide a regular expression which enables the lint to
-        // target custom reactive hooks.
-        let name;
-        try {
-          name = toPropertyAccessString(node);
-        } catch (error) {
-          if (/Unsupported node type/.test(error.message)) {
-            return 0;
-          } else {
-            throw error;
-          }
-        }
-        return options.additionalHooks.test(name) ? 0 : -1;
-      } else {
-        return -1;
-      }
-  }
-}
-
-/**
- * ESLint won't assign node.parent to references from context.getScope()
- *
- * So instead we search for the node from an ancestor assigning node.parent
- * as we go. This mutates the AST.
- *
- * This traversal is:
- * - optimized by only searching nodes with a range surrounding our target node
- * - agnostic to AST node types, it looks for `{ type: string, ... }`
- */
-function fastFindReferenceWithParent(start, target) {
-  const queue = [start];
-  let item = null;
-
-  while (queue.length) {
-    item = queue.shift();
-
-    if (isSameIdentifier(item, target)) {
-      return item;
-    }
-
-    if (!isAncestorNodeOf(item, target)) {
-      continue;
-    }
-
-    for (const [key, value] of Object.entries(item)) {
-      if (key === 'parent') {
-        continue;
-      }
-      if (isNodeLike(value)) {
-        value.parent = item;
-        queue.push(value);
-      } else if (Array.isArray(value)) {
-        value.forEach(val => {
-          if (isNodeLike(val)) {
-            val.parent = item;
-            queue.push(val);
-          }
-        });
-      }
-    }
-  }
-
-  return null;
-}
-
-function joinEnglish(arr) {
-  let s = '';
-  for (let i = 0; i < arr.length; i++) {
-    s += arr[i];
-    if (i === 0 && arr.length === 2) {
-      s += ' and ';
-    } else if (i === arr.length - 2 && arr.length > 2) {
-      s += ', and ';
-    } else if (i < arr.length - 1) {
-      s += ', ';
-    }
-  }
-  return s;
-}
-
-function isNodeLike(val) {
-  return (
-    typeof val === 'object' &&
-    val !== null &&
-    !Array.isArray(val) &&
-    typeof val.type === 'string'
-  );
-}
-
-function isSameIdentifier(a, b) {
-  return (
-    a.type === 'Identifier' &&
-    a.name === b.name &&
-    a.range[0] === b.range[0] &&
-    a.range[1] === b.range[1]
-  );
-}
-
-function isAncestorNodeOf(a, b) {
-  return a.range[0] <= b.range[0] && a.range[1] >= b.range[1];
-}
